requirements.txt
Pillow>=8.2.0
pycryptodome>=3.9.9
